test: migrate range_authorship.rs unit tests to integration tests
Migrated all 28 tests from src/authorship/range_authorship.rs to tests/integration/range_authorship_unit.rs, converting from TmpRepo to TestRepo.

Changes:
- Made EMPTY_TREE_HASH pub in src/authorship/range_authorship.rs
- Created tests/integration/range_authorship_unit.rs with all 28 tests:
  - 11 range_authorship tests (TmpRepo -> TestRepo conversion)
  - 17 should_ignore_file pure tests (direct function usage)
- Added range_authorship_unit module to tests/integration/main.rs
- Removed #[cfg(test)] mod tests from src/authorship/range_authorship.rs

All tests pass: cargo test --test integration range_authorship_unit
